back_end/requirements.txt
azure_storage==0.36.0
face_recognition==1.3.0
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.23.4
opencv_python_headless==4.6.0.66
pymongo==4.3.2
gunicorn==20.1.0
azure-storage-blob==12.14.0